movie_site/requirements.txt
Django
djangorestframework
psycopg2-binary
six
flake8
drf-extra-fields
django-filter
django-environ
Pillow